moved dispatcher to ts
