--- frontend/onied/src/components/blocks/blockViewContainer.jsx
@@ -1,5 +1,0 @@
-import classes from "./blocks.module.css";
-
-export default function BlockViewContainer({ children }) {
-  return <div className={classes.blockViewContainer}>{children}</div>;
-}
