+++ frontend/onied/src/components/blocks/blockViewContainer.tsx
@@ -1,0 +1,6 @@
+import { PropsWithChildren } from "react";
+import classes from "./blocks.module.css";
+
+export default function BlockViewContainer({ children }: PropsWithChildren) {
+  return <div className={classes.blockViewContainer}>{children}</div>;
+}
